refactor: convert authenticate file from JavaScript to TypeScript
